Move clippy lints from `flake.nix` to `Cargo.toml`
This way running `cargo clippy` on the command line uses the same lints.
A small downside is that the `Cargo.toml` file feeds into the Nix build, so if I change any of these lints or add new ones, Nix recompiles everything.
A solution to this may be to teach crane to filter this file before using it to build the package and only use the relevant part while running clippy.