Add $ORIGIN directive in test zone file
According to https://www.zytrax.com/books/dns/ch8/origin.html:
"If an $ORIGIN directive is not defined - BIND synthesizes an $ORIGIN from the zone name in the named.conf file"
So the test worked without the $ORIGIN directive but it's good practice and less confusing to always add it.

While we're at it, also reduce code duplication by using the $ORIGIN directive to our advantage.